pyproject.toml
[project]
name = "bill-management-rag-mongo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chroma>=0.2.0",
    "fastapi[standard]>=0.128.0",
    "ipykernel>=7.1.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "pdf2image>=1.17.0",
    "pillow>=12.1.0",
    "pinecone>=8.0.0",
    "pinecone-client>=6.0.0",
    "pydantic>=2.12.5",
    "pymongo>=4.15.5",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sentence-transformers>=5.2.0",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
    "uuid>=1.30",
]
